+++ map-script.js
@@ -109,7 +109,10 @@ function initializeMap() {
         // Load Power Stations
         var powerStationsUrl = 'geojson/power_stations.geojson';
         map.loadImage("images/power-stations.png", function(error, image) { //this is where we load the image file
-        if (error) throw error;
+        if (error) {
+            showMapTokenWarning('A map marker image could not be loaded. Check the local image assets and reload the page.');
+            return;
+        }
             map.addImage("custom-marker", image);
             map.addLayer({
             'id': 'power_stations',
@@ -129,7 +132,10 @@ function initializeMap() {
         // Load Cell Towers
         var cellTowersUrl = 'geojson/cell_towers.geojson';
         map.loadImage("images/cell-towers.png", function(error, image) { //this is where we load the image file
-        if (error) throw error;
+        if (error) {
+            showMapTokenWarning('A map marker image could not be loaded. Check the local image assets and reload the page.');
+            return;
+        }
             map.addImage("cell-tower", image);
             map.addLayer({
             'id': 'cell_towers',